double

+++ src/tungsten-test/dft-realization-gen.cpp
@@ -113,14 +113,14 @@ void rational_quadratic_sphere() {
 
     for (int num_reals = 0; num_reals < 500; num_reals++) {
         for (int i = 0; i < Fcov.size(); i++) {
-            Vec2f u = gp.rand_normal_2(sampler);
+            Vec2d u = gp.rand_normal_2(sampler);
             Fcov_sample[i] = sqrt(Fcov[i] / std::complex<double>(cov.size())) * (std::complex<double>(u.x(), u.y()));
         }
 
         fftw_execute(plan);
 
         for (int i = 0; i < Fcov.size(); i++) {
-            Vec2f u = gp.rand_normal_2(sampler);
+            Vec2d u = gp.rand_normal_2(sampler);
             real[i] += std::complex<double>((*gp._mean)(derivs[i], points[i], Vec3f(0.f)));
         }
 
@@ -132,7 +132,7 @@ void rational_quadratic_sphere() {
     }
 
     for (int i = 0; i < Fcov.size(); i++) {
-        Vec2f u = gp.rand_normal_2(sampler);
+        Vec2d u = gp.rand_normal_2(sampler);
         real[i] = std::complex<double>((*gp._mean)(derivs[i], points[i], Vec3f(0.f)));
     }
 
@@ -188,14 +188,14 @@ void rational_quadratic_sphere_3D() {
             openvdb::io::File file(tinyformat::format("3d-reals/%s-sample%d.vdb", gp._cov->id(), num_reals));
 
             for (int i = 0; i < Fcov.size(); i++) {
-                Vec2f u = gp.rand_normal_2(sampler);
+                Vec2d u = gp.rand_normal_2(sampler);
                 Fcov_sample[i] = sqrt(Fcov[i] / std::complex<double>(4*cov.size())) * (std::complex<double>(u.x(), u.y()));
             }
 
             fftw_execute(plan);
 
             for (int i = 0; i < Fcov.size(); i++) {
-                Vec2f u = gp.rand_normal_2(sampler);
+                Vec2d u = gp.rand_normal_2(sampler);
                 real[i] += std::complex<double>((*gp._mean)(derivs[i], points[i], Vec3f(0.f)));
             }
 
@@ -236,7 +236,7 @@ void rational_quadratic_sphere_3D() {
     }
 
     for (int i = 0; i < Fcov.size(); i++) {
-        Vec2f u = gp.rand_normal_2(sampler);
+        Vec2d u = gp.rand_normal_2(sampler);
         real[i] = std::complex<double>((*gp._mean)(derivs[i], points[i], Vec3f(0.f)));
     }
